@@ -6,18 +6,23 @@
     "test": "karma start ClientApp/test/karma.conf.js"
   },
   "dependencies": {
-    "@angular/animations": "4.2.5",
-    "@angular/common": "4.2.5",
-    "@angular/compiler": "^4.2.5",
-    "@angular/compiler-cli": "^4.4.7",
-    "@angular/core": "^4.2.5",
-    "@angular/forms": "4.2.5",
-    "@angular/http": "4.2.5",
-    "@angular/platform-browser": "4.2.5",
-    "@angular/platform-browser-dynamic": "4.2.5",
-    "@angular/platform-server": "4.2.5",
-    "@angular/router": "4.2.5",
-    "@ngtools/webpack": "^1.5.0",
+    "@angular/animations": "6.0.9",
+    "@angular/cli": "^6.0.8",
+    "@angular/common": "6.0.9",
+    "@angular/compiler": "^6.0.9",
+    "@angular/compiler-cli": "^6.0.9",
+    "@angular/core": "^6.0.9",
+    "@angular/forms": "6.0.9",
+    "@angular/http": "6.0.9",
+    "@angular/platform-browser": "6.0.9",
+    "@angular/platform-browser-dynamic": "6.0.9",
+    "@angular/platform-server": "6.0.9",
+    "@angular/router": "6.0.9",
+    "@fortawesome/angular-fontawesome": "^0.1.1",
+    "@fortawesome/fontawesome-svg-core": "^1.2.0",
+    "@fortawesome/free-solid-svg-icons": "^5.1.0",
+    "@ng-bootstrap/ng-bootstrap": "^2.2.0",
+    "@ngtools/webpack": "^1.10.2",
     "@types/chai": "4.0.1",
     "@types/jasmine": "2.5.53",
     "@types/webpack-env": "1.13.0",
@@ -48,17 +53,19 @@
     "karma-webpack": "2.0.3",
     "mydatepicker": "^2.6.3",
     "ng2-dragula": "^1.5.0",
+    "ngx-bootstrap": "^3.0.1",
     "preboot": "4.5.2",
     "raw-loader": "0.5.1",
     "reflect-metadata": "0.1.10",
-    "rxjs": "^5.4.2",
+    "rxjs": "^6.2.1",
+    "rxjs-compat": "^6.0.0-rc.0",
     "style-loader": "0.18.2",
     "to-string-loader": "1.1.5",
-    "typescript": "2.4.1",
+    "typescript": "2.7.2",
     "url-loader": "0.5.9",
     "webpack": "^2.5.1",
     "webpack-hot-middleware": "2.18.2",
     "webpack-merge": "^4.1.3",
-    "zone.js": "0.8.12"
+    "zone.js": "0.8.26"
   }
 }
