@@ -4,12 +4,17 @@ import { AppModuleShared } from './app.shared.module';
 import { LayoutComponent } from './modules/layout/pages/layout.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
+import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
+import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
+
 @NgModule({
     bootstrap: [ LayoutComponent ],
     imports: [
         BrowserModule,
         BrowserAnimationsModule,
-        AppModuleShared
+        AppModuleShared,
+        NgbModule.forRoot(),
+        FontAwesomeModule
     ],
     providers: [
         { provide: 'BASE_URL', useFactory: getBaseUrl }
