@@ -2,8 +2,7 @@ import { Component } from '@angular/core';
 
 @Component({
     selector: 'content',
-    templateUrl: './content.component.html',
-    styleUrls: ['./content.component.css']
+    templateUrl: './content.component.html'
 })
 export class ContentComponent {
 }
