@@ -1,18 +1,11 @@
-import { Component } from '@angular/core';
-import { IMyDpOptions } from 'mydatepicker';
-import { MyDatePicker } from 'mydatepicker';
+import { Component, ViewEncapsulation } from '@angular/core';
 
 @Component({
     selector: 'daypicker',
     templateUrl: './daypicker.component.html',
-    styleUrls: ['./daypicker.component.css']
+    styleUrls: ['./daypicker.component.css'],
+    encapsulation: ViewEncapsulation.None
 })
-export class DayPickerComponent extends MyDatePicker {
-
-    public myDatePickerOptions: IMyDpOptions = {
-        dateFormat: 'dd.mm.yyyy',
-    };
-    public model: any = {
-        date: { year: 2018, month: 10, day: 9 }
-    };
+export class DayPickerComponent {
+   
 }
