@@ -3,18 +3,23 @@
 <head>
     <meta name="viewport" content="width=device-width, initial-scale=1.0" charset="utf-8">
     <title>@ViewData["Title"] - TasksAngular</title>
+
     <base href="~/" />
+    <link href="~/normalize.css" rel="stylesheet" />
+
+    <link rel="stylesheet" href="~/dist/vendor.css" asp-append-version="true" />
+
     <link href="https://fonts.googleapis.com/css?family=Copse|Scope+One|Spectral+SC" rel="stylesheet">
     <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
     <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto:300,400,500,700,400italic">
-    <link rel="stylesheet" href="~/dist/vendor.css" asp-append-version="true" />
-    <link href="~/normalize.css" rel="stylesheet" />
-    <link href="~/site.css" rel="stylesheet" />
 
+    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.1/css/bootstrap.min.css" integrity="sha384-WskhaSGFgHYWDcbwN70/dfYBj47jz9qbsMId/iRN3ewGhXQFZCSftd1LZCfmhktB" crossorigin="anonymous">
+    <link rel="stylesheet" type="text/css" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css" />
+
+    <link href="~/site.css" rel="stylesheet" />
 </head>
 <body>
     @RenderBody()
-
     @RenderSection("scripts", required: false)
 </body>
 </html>
